♻️[#78] Refactor: 캐러셀 컴포넌트 구조 분리
